requirements.txt
Django>=4.2
django-ninja
PyJWT
django-redis
redis
